PHOENIX-3452 Enable mode of payment bank challan

+++ egov/egov-collectionweb/src/main/java/org/egov/collection/web/actions/receipts/MiscellaneousFileUploadAction.java
@@ -183,7 +183,7 @@ public class MiscellaneousFileUploadAction extends BaseFormAction {
                     + "UPL");
 
             final Bankaccount account = (Bankaccount) persistenceService.find("from Bankaccount where accountnumber=?", input[4]);
-            receiptAction.setBankAccountId(account.getId().intValue());
+            receiptAction.setBankAccountId(account.getId());
 
             receiptAction.setInstrHeaderBank(instrHeaderBank);
             receiptAction.setInstrumentTypeCashOrCard(CollectionConstants.INSTRUMENTTYPE_BANK);
